User tests

+++ routes/api.php
@@ -19,11 +19,11 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 });
 
 Route::post("/user", ['App\Http\Controllers\UserController', "cadastrar"]);
-Route::get("/user", ['App\Http\Controllers\UserController', "listar"])->middleware("jwt");
-Route::get("/user/{uuid}", ['App\Http\Controllers\UserController', "buscar"]);
+Route::get("/user", ['App\Http\Controllers\UserController', "listar"])->middleware("jwt:admin");
+Route::get("/user/{uuid}", ['App\Http\Controllers\UserController', "buscar"])->middleware("jwt:self");
 Route::put("/user/{uuid}", ['App\Http\Controllers\UserController', "atualizar"]);
-Route::patch("/user/{uuid}", ['App\Http\Controllers\UserController', "admin"]);
-Route::delete("/user/{uuid}", ['App\Http\Controllers\UserController', "remover"])->middleware("jwt");
+Route::patch("/user/{uuid}/admin", ['App\Http\Controllers\UserController', "admin"]);
+Route::delete("/user/{uuid}", ['App\Http\Controllers\UserController', "remover"])->middleware("jwt:admin");
 
 Route::post("/login", ["App\Http\Controllers\LoginController", "login"]);
 Route::post("/logout", ["App\Http\Controllers\LoginController", "logout"]);
